typos

+++ Model/src/main/java/org/example/model/statements/FirstTypeSummary.java
@@ -14,8 +14,8 @@ public class FirstTypeSummary extends Summary {
 
     @Override
     protected double getDegreeOfCovering(List<Entry> entries) {
-        int m = entries.size();
-        return ((double) entries
+        double m = entries.size();
+        return (entries
                 .stream()
                 .filter(e -> this.summarizer.getMembershipFunctionValueFor(e) > 0)
                 .count()) / m;
